Reference to NeuroImage corrected
