+++ Checklist_v2.md
@@ -1,0 +1,13 @@
+# Check list 
+
+Make sure these points are fixed before release of v2.0.0 of NetDynFlow.
+
+1. Choose one version for the core module between *core1.py* and *core2.py*.
+2. Clean imports from `__init__.py` file.
+3. Correctly cite the published papers everywhere ( README.md, `__init__.py`, etc. )
+4. Simplify internal variable naming. Homogenize internal naming with pyMOU package.
+5. Add a folder with tutorial Jupyter Notebooks.
+
+
+
+
