+++ NamingConventions.md
@@ -1,0 +1,15 @@
+## Internal variable naming conventions
+
+
+<br/>
+
+- con_matrix --> C
+- tau_const --> tau
+- sigma_mat --> sigma
+- n_nodes --> N
+- GenerateTensors() --> ??
+
+
+
+
+
