+++ netdynflow/core.py
@@ -26,6 +26,8 @@ JacobianMOU
 
 Generation of main tensors
 --------------------------
+GenerateTensors
+    Generic function to create time evolution of communicability or flow.
 DynCom
     Returns the temporal evolution of a network's dynamic communicability.
 DynFlow
@@ -39,9 +41,9 @@ Reference and Citation
 ----------------------
 1. M. Gilson, N. Kouvaris, G. Deco & G.Zamora-Lopez "Framework based on communi-
 cability and flow to analyze complex networks" Phys. Rev. E 97, 052301 (2018).
-2. M. Gilson, N. Kouvaris, et al. "Analysis of brain network dynamics estimated
-from fMRI data: A new framework based on communicability and flow"
-bioRxiv (2018). DOI: https://doi.org/10.1101/421883.
+2. M. Gilson, N. Kouvaris, et al. "Network analysis of whole-brain fMRI
+dynamics: A new framework based on dynamic communicability" NeuroImage 201,
+116007 (2019).
 
 
 ...moduleauthor:: Gorka Zamora-Lopez <galib@zamora-lopez.xyz>
@@ -53,7 +55,7 @@ import numpy as np
 import numpy.linalg
 import scipy.linalg
 
-# __all__ = ['JaccobianMOU', 'DynCom', 'DynFlow', 'FullFlow', 'IntrinsicFlow']
+__all__ = ['JaccobianMOU', 'DynCom', 'DynFlow', 'FullFlow', 'IntrinsicFlow']
 
 
 ## USEFUL FUNCTIONS ##########################################################
@@ -106,6 +108,98 @@ def JacobianMOU(con_matrix, tau_const):
 
 
 ## GENERATION OF THE MAIN TENSORS #############################################
+def GenerateTensors(con_matrix, tau_const, sigma_mat, tmax=20, timestep=0.1,
+                    normed=True, case='DynCom'):
+    ## GORKA: Function probably needs another name.
+    """Generic function to create time evolution of communicability or flow.
+
+    Parameters
+    ----------
+    con_matrix : ndarray of rank-2
+        The adjacency matrix of the network.
+    tau_const : real valued number, or ndarray of rank-1
+        The decay rate at the nodes. Positive value expected.
+        If a number is given, then the function considers all nodes have same
+        decay rate. Alternatively, an array can be inputed with the decay rate
+        of each node.
+    sigma_mat : ndarray of rank-2
+        The covariance matrix of fluctuating inputs.
+    tmax : real valued number, positive (optional)
+        Final time for integration.
+    timestep : real valued number, positive (optional)
+        Sampling time-step.
+        Warning: Not an integration step, just the desired sampling rate.
+    normed : boolean (optional)
+        If True, normalises the tensor by the scaling factor, to make networks
+        of different size comparable.
+
+    Returns
+    -------
+    flow_tensor : ndarray of rank-3
+        Temporal evolution of the network's dynamic flow or communicability.
+        A tensor of shape (tmax*timestep) x n_nodes x n_nodes, where n_nodes is
+        the number of nodes.
+    """
+    # 0) SECURITY CHECKS
+    caselist = ['DynCom', 'DynFlow', 'FullFlow', 'IntrinsicFlow']
+    if case not in caselist:
+        raise ValueError( "Please enter one of accepted cases: %s" %str(caselist) )
+
+    if tmax <= 0.0: raise ValueError("'tmax' must be positive")
+    if timestep <= 0.0: raise ValueError( "'timestep' must be positive")
+    if timestep > tmax: raise ValueError("Incompatible values, timestep < tmax given")
+
+    # 1) CALCULATE THE JACOBIAN MATRIX
+    jacobian = JacobianMOU(con_matrix, tau_const)
+    jacobian_diag = np.diagonal(jacobian)
+    n_nodes = len(jacobian)
+
+    # 2) CALCULATE THE DYNAMIC FLOW
+    # 2.1) Calculate the extrinsic flow over integration time
+    n_t = int(tmax / timestep) + 1
+    sigma_sqrt_mat = scipy.linalg.sqrtm(sigma_mat)
+
+    flow_tensor = np.zeros((n_t,n_nodes,n_nodes), dtype=np.float)
+    if case == 'DynCom':
+        for i_t in range(n_t):
+            t = i_t * timestep
+            # Calculate the term for jacobian_diag without using expm(), to speed up
+            jacobian_diag_t = np.diag( np.exp(jacobian_diag * t) )
+            # Calculate the dynamic communicability at time t
+            flow_tensor[i_t] = scipy.linalg.expm(jacobian * t) - jacobian_diag_t
+
+    elif case == 'DynFlow':
+        for i_t in range(n_t):
+            t = i_t * timestep
+            # Calculate the term for jacobian_diag without using expm(), to speed up
+            jacobian_diag_t = np.diag( np.exp(jacobian_diag * t) )
+            # Calculate the dynamic communicability at time t
+            flow_tensor[i_t] = np.dot( sigma_mat, \
+                            (scipy.linalg.expm(jacobian * t) - jacobian_diag_t) )
+
+    elif case == 'IntrinsicFlow':
+        for i_t in range(n_t):
+            t = i_t * timestep
+            # Calculate the term for jacobian_diag without using expm(), to speed up
+            jacobian_diag_t = np.diag( np.exp(jacobian_diag * t) )
+            # Calculate the dynamic communicability at time t.
+            flow_tensor[i_t] = np.dot( sigma_mat, jacobian_diag_t)
+
+    elif case == 'FullFlow':
+        for i_t in range(n_t):
+            t = i_t * timestep
+            # Calculate the non-normalised flow at time t.
+            flow_tensor[i_t] = np.dot( sigma_mat, scipy.linalg.expm(jacobian * t) )
+
+    # 2.2) Normalise by the scaling factor
+    if normed:
+        scaling_factor = (-1./jacobian_diag).sum()
+        flow_tensor /= scaling_factor
+
+    return flow_tensor
+
+
+## Wrappers using GenerateTensors() ___________________________________________
 def DynFlow(con_matrix, tau_const, sigma_mat, tmax=20, timestep=0.1, normed=True):
     """Returns the extrinsinc flow on a network over time for a given input.
 
@@ -131,40 +225,14 @@ def DynFlow(con_matrix, tau_const, sigma_mat, tmax=20, timestep=0.1, normed=True
 
     Returns
     -------
-    flow_tensor : ndarray of rank-3
+    dynflow_tensor : ndarray of rank-3
         Temporal evolution of the network's dynamic communicability. A tensor
         of shape (tmax*timestep) x n_nodes x n_nodes, where n_nodes is the number of nodes.
     """
-    # 0) SECURITY CHECKS
-    if tmax <= 0.0: raise ValueError("'tmax' must be positive")
-    if timestep <= 0.0: raise ValueError( "'timestep' must be positive")
-    if timestep > tmax: raise ValueError("Incompatible values, timestep < tmax given")
+    dynflow_tensor = GenerateTensors(con_matrix, tau_const, sigma_mat, tmax=tmax,
+                    timestep=timestep, normed=normed, case='DynFlow')
 
-    # 1) CALCULATE THE JACOBIAN MATRIX
-    jacobian = JacobianMOU(con_matrix, tau_const)
-    jacobian_diag = np.diagonal(jacobian)
-    n_nodes = len(jacobian)
-
-    # 2) CALCULATE THE DYNAMIC FLOW
-    # 2.1) Calculate the extrinsic flow over integration time
-    n_t = int(tmax / timestep) + 1
-    sigma_sqrt_mat = scipy.linalg.sqrtm(sigma_mat)
-
-    flow_tensor = np.zeros((n_t,n_nodes,n_nodes), dtype=np.float)
-    for i_t in range(n_t):
-        t = i_t * timestep
-        # Calculate the term for jacobian_diag without using expm(), to speed up
-        jacobian_diag_t = np.diag( np.exp(jacobian_diag * t) )
-        # Calculate the dynamic communicability at time t
-        flow_tensor[i_t] = np.dot( sigma_mat, \
-                        (scipy.linalg.expm(jacobian * t) - jacobian_diag_t) )
-
-    # 2.2) Normalise by the scaling factor
-    if normed:
-        scaling_factor = (-1./jacobian_diag).sum()
-        flow_tensor /= scaling_factor
-
-    return flow_tensor
+    return dynflow_tensor
 
 def DynCom(con_matrix, tau_const, tmax=20, timestep=0.1, normed=True):
     """Returns the temporal evolution of a network's dynamic communicability.
@@ -193,32 +261,10 @@ def DynCom(con_matrix, tau_const, tmax=20, timestep=0.1, normed=True):
         Temporal evolution of the network's dynamic communicability. A tensor
         of shape (tmax*timestep) x N x N, where N is the number of nodes.
     """
-    # 0) SECURITY CHECKS
-    if tmax <= 0.0: raise ValueError("'tmax' must be positive")
-    if timestep <= 0.0: raise ValueError( "'timestep' must be positive")
-    if timestep > tmax: raise ValueError("Incompatible values, timestep < tmax given")
-
-    # 1) CALCULATE THE JACOBIAN MATRIX
-    jacobian = JacobianMOU(con_matrix, tau_const)
-    jacobian_diag = np.diagonal(jacobian)
-    n_nodes = len(jacobian)
-
-    # 2) CALCULATE THE DYNAMIC COMMUNICABILITY
-    # 2.1) Calculate the extrinsic flow over integration time
-    n_t = int(tmax / timestep) + 1
-
-    dyncom_tensor = np.zeros((n_t,n_nodes,n_nodes), dtype=np.float)
-    for i_t in range(n_t):
-        t = i_t * timestep
-        # Calculate the term for jacobian_diag without using expm(), to speed up
-        jacobian_diag_t = np.diag( np.exp(jacobian_diag * t) )
-        # Calculate the dynamic communicability at time t
-        dyncom_tensor[i_t] = scipy.linalg.expm(jacobian * t) - jacobian_diag_t
-
-    # 2.2) Normalise by the scaling factor
-    if normed:
-        scaling_factor = (-1./jacobian_diag).sum()
-        dyncom_tensor /= scaling_factor
+    n_nodes = len(con_matrix)
+    sigma_mat = np.identity(n_nodes, dtype=np.float)
+    dyncom_tensor = GenerateTensors(con_matrix, tau_const, sigma_mat, tmax=tmax,
+                    timestep=timestep, normed=normed, case='DynCom')
 
     return dyncom_tensor
 
@@ -251,33 +297,8 @@ def IntrinsicFlow(con_matrix, tau_const, sigma_mat, tmax=20, timestep=0.1, norme
         Temporal evolution of the network's dynamic communicability. A tensor
         of shape (tmax*timestep) x N x N, where N is the number of nodes.
     """
-    # 0) SECURITY CHECKS
-    if tmax <= 0.0: raise ValueError("'tmax' must be positive")
-    if timestep <= 0.0: raise ValueError( "'timestep' must be positive")
-    if timestep > tmax: raise ValueError("Incompatible values, timestep < tmax given")
-
-    # 1) CALCULATE THE JACOBIAN MATRIX
-    jacobian = JacobianMOU(con_matrix, tau_const)
-    jacobian_diag = np.diagonal(jacobian)
-    n_nodes = len(jacobian)
-
-    # 2) CALCULATE THE DYNAMIC COMMUNICABILITY
-    # 2.1) Calculate the extrinsic flow over integration time
-    n_t = int(tmax / timestep) + 1
-    sigma_sqrt_mat = scipy.linalg.sqrtm(sigma_mat)
-
-    flow_tensor = np.zeros((n_t,n_nodes,n_nodes), dtype=np.float)
-    for i_t in range(n_t):
-        t = i_t * timestep
-        # Calculate the term for jacobian_diag without using expm(), to speed up
-        jacobian_diag_t = np.diag( np.exp(jacobian_diag * t) )
-        # Calculate the dynamic communicability at time t.
-        flow_tensor[i_t] = np.dot( sigma_mat, jacobian_diag_t)
-
-    # 1.3) Normalise by the scaling factor
-    if normed:
-        scaling_factor = (-1./jacobian_diag).sum()
-        flow_tensor /= scaling_factor
+    flow_tensor = GenerateTensors(con_matrix, tau_const, sigma_mat, tmax=tmax,
+                    timestep=timestep, normed=normed, case='IntrinsicFlow')
 
     return flow_tensor
 
@@ -309,45 +330,10 @@ def FullFlow(con_matrix, tau_const, sigma_mat, tmax=20, timestep=0.1, normed=Tru
         Temporal evolution of the network's flow. A tensor of shape
         (tmax*timestep) x N x N, where N is the number of nodes.
     """
-    # 0) SECURITY CHECKS
-    if tmax <= 0.0: raise ValueError("'tmax' must be positive")
-    if timestep <= 0.0: raise ValueError( "'timestep' must be positive")
-    if timestep > tmax: raise ValueError("Incompatible values, timestep < tmax given")
-
-    # 1) CALCULATE THE JACOBIAN MATRIX
-    jacobian = JacobianMOU(con_matrix, tau_const)
-    jacobian_diag = np.diagonal(jacobian)
-    n_nodes = len(jacobian)
-
-    # 2) CALCULATE THE DYNAMIC FLOW
-    # 2.1) Calculate the extrinsic flow over integration time
-    n_t = int(tmax / timestep) + 1
-    sigma_sqrt_mat = scipy.linalg.sqrtm(sigma_mat)
-
-    flow_tensor = np.zeros((n_t,n_nodes,n_nodes), dtype=np.float)
-    for i_t in range(n_t):
-        t = i_t * timestep
-        # Calculate the non-normalised flow at time t.
-        flow_tensor[i_t] = np.dot( sigma_mat, scipy.linalg.expm(jacobian * t) )
-
-    # 2.2) Normalise by the scaling factor
-    if normed:
-        scaling_factor = (-1./jacobian_diag).sum()
-        flow_tensor /= scaling_factor
+    flow_tensor = GenerateTensors(con_matrix, tau_const, sigma_mat, tmax=tmax,
+                            timestep=timestep, normed=normed, case='FullFlow')
 
     return flow_tensor
-
-## GORKA: Nice try, but this one seems very inefficient to me.
-## It iterates twice through time.
-# def FullFlow(con_matrix, tau_const, sigma_mat, tmax=20, timestep=0.1, normed=True):
-#     """Returns the complete flow on a network over time for a given input.
-#
-#     ...
-#     """
-#     flow_tensor = DynFlow(con_matrix, tau_const, np.eye(n_nodes), tmax, timestep, normed) \
-#                + IntrinsicFlow(con_matrix, tau_const, np.eye(n_nodes), tmax, timestep, normed)
-#
-#     return flow_tensor
 
 
 ##
