+++ netdynflow/metrics.py
@@ -28,9 +28,9 @@ Reference and Citation
 ----------------------
 1. M. Gilson, N. Kouvaris, G. Deco & G.Zamora-Lopez "Framework based on communi-
 cability and flow to analyze complex networks" Phys. Rev. E 97, 052301 (2018).
-2. M. Gilson, N. Kouvaris, et al. "Analysis of brain network dynamics estimated
-from fMRI data: A new framework based on communicability and flow"
-bioRxiv (2018). DOI: https://doi.org/10.1101/421883.
+2. M. Gilson, N. Kouvaris, et al. "Network analysis of whole-brain fMRI
+dynamics: A new framework based on dynamic communicability" NeuroImage 201,
+116007 (2019).
 
 
 ...moduleauthor:: Gorka Zamora-Lopez <galib@zamora-lopez.xyz>
